backend/django_Admin3/requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.3
black==26.1.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.13.1
css-inline==0.19.0
cssselect==1.3.0
cssutils==2.11.1
Django==6.0.1
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-email-templates==0.1
django-environ==0.12.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
docutils==0.22.4
dotmap==1.3.30
drf-yasg==1.21.14
et_xmlfile==2.0.0
flake8==7.3.0
fuzzywuzzy==0.18.0
gunicorn==24.0.0
html2text==2025.4.15
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
Jinja2==3.1.6
Levenshtein==0.27.3
lxml==6.0.2
MarkupSafe==3.0.3
mccabe==0.7.0
mjml-python==1.3.7
more-itertools==10.8.0
mypy_extensions==1.1.0
numpy==2.4.1
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
pathspec==1.0.3
platformdirs==4.5.1
pluggy==1.6.0
premailer==3.10.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-Levenshtein==0.27.3
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
dj-database-url==3.1.0
django-anymail[sendgrid]==14.0
jsonschema==4.26.0
whitenoise==6.11.0
pact-python==2.2.2
